+++ lib/pio/open_flow10/enqueue.rb
@@ -1,7 +1,7 @@
 require 'bindata'
 require 'forwardable'
 require 'pio/monkey_patch/integer'
-require 'pio/open_flow/port_number'
+require 'pio/open_flow10/port16'
 
 module Pio
   # An action to enqueue the packet on the specified queue attached to
@@ -11,9 +11,9 @@ module Pio
     class Format < BinData::Record
       endian :big
 
-      uint16 :type, value: 11
-      uint16 :message_length, value: 16
-      port_number :port_number
+      uint16 :action_type, value: 11
+      uint16 :action_length, value: 16
+      port16 :port
       uint48 :padding
       hide :padding
       uint32 :queue_id
@@ -27,28 +27,28 @@ module Pio
 
     extend Forwardable
 
-    def_delegators :@format, :type
-    def_delegators :@format, :message_length
-    def_delegators :@format, :port_number
+    def_delegators :@format, :action_type
+    def_delegator :@format, :action_length, :length
+    def_delegators :@format, :port
     def_delegators :@format, :queue_id
     def_delegator :@format, :to_binary_s, :to_binary
 
     def initialize(user_options)
-      validate_port_number user_options
+      validate_port user_options
       validate_queue_id user_options
       @format = Format.new(user_options)
     end
 
     private
 
-    def validate_port_number(user_options)
-      port_number = user_options.fetch(:port_number)
-      if port_number.is_a?(Symbol) && port_number != :in_port
+    def validate_port(user_options)
+      port = user_options.fetch(:port)
+      if port.is_a?(Symbol) && port != :in_port
         fail(ArgumentError,
-             ':port_number must be a valid physical port or :in_port')
+             ':port must be a valid physical port or :in_port')
       end
     rescue KeyError
-      raise ArgumentError, ':port_number is a mandatory option'
+      raise ArgumentError, ':port is a mandatory option'
     end
 
     def validate_queue_id(user_options)