+++ lib/pio/open_flow10/features.rb
@@ -1,12 +1,13 @@
 require 'pio/open_flow'
+require 'pio/open_flow10/phy_port16'
 
 module Pio
   module OpenFlow10
-    # OpenFlow 1.0 Features Request and Reply message.
+    # Features Request and Reply message.
     class Features
-      # OpenFlow 1.0 Features Request message.
+      # Features Request message.
       class Request < OpenFlow::Message
-        # OpenFlow 1.0 Features Request message
+        # Features Request message format.
         class Format < BinData::Record
           extend OpenFlow::Format
 
@@ -16,11 +17,6 @@ module Pio
           def user_data
             body
           end
-        end
-
-        def initialize(user_options = {})
-          validate_user_options user_options
-          @format = Format.new(header: parse_header_options(user_options))
         end
       end
 
@@ -65,7 +61,7 @@ module Pio
           hide :padding
           capabilities :capabilities
           actions_flag :actions
-          array :ports, type: :phy_port, read_until: :eof
+          array :ports, type: :phy_port16, read_until: :eof
 
           def dpid
             datapath_id
@@ -96,7 +92,7 @@ module Pio
 
           def physical_ports
             ports.select do |each|
-              each.port_no <= PortNumber::MAX
+              each.port_no <= Port16::MAX
             end
           end
         end