+++ lib/pio/open_flow10/set_ip_tos.rb
@@ -9,8 +9,8 @@ module Pio
     class Format < BinData::Record
       endian :big
 
-      uint16 :type, value: 8
-      uint16 :message_length, value: 8
+      uint16 :action_type, value: 8
+      uint16 :action_length, value: 8
       uint8 :type_of_service
       uint24 :padding
       hide :padding
@@ -24,8 +24,8 @@ module Pio
 
     extend Forwardable
 
-    def_delegators :@format, :type
-    def_delegators :@format, :message_length
+    def_delegators :@format, :action_type
+    def_delegator :@format, :action_length, :length
     def_delegators :@format, :type_of_service
     def_delegator :@format, :to_binary_s, :to_binary