+++ lib/pio/open_flow10/strip_vlan_header.rb
@@ -8,22 +8,22 @@ module Pio
     class Format < BinData::Record
       endian :big
 
-      uint16 :type, value: 3
-      uint16 :message_length, value: 8
+      uint16 :action_type, value: 3
+      uint16 :action_length, value: 8
       uint32 :padding
       hide :padding
     end
 
     def self.read(raw_data)
-      strip_vlan = allocate
-      strip_vlan.instance_variable_set :@format, Format.read(raw_data)
-      strip_vlan
+      allocate.tap do |strip_vlan|
+        strip_vlan.instance_variable_set :@format, Format.read(raw_data)
+      end
     end
 
     extend Forwardable
 
-    def_delegators :@format, :type
-    def_delegators :@format, :message_length
+    def_delegators :@format, :action_type
+    def_delegator :@format, :action_length, :length
     def_delegator :@format, :to_binary_s, :to_binary
 
     def initialize