+++ spec/pio/open_flow10/phy_port16_spec.rb
@@ -1,14 +1,14 @@
 require 'pio'
 
-describe Pio::OpenFlow::PhyPort do
+describe Pio::OpenFlow10::PhyPort16 do
   describe '.new' do
     When(:phy_port) do
-      Pio::OpenFlow::PhyPort.new(port_no: 1,
-                                 hardware_address: '11:22:33:44:55:66',
-                                 name: 'port123',
-                                 config: [:port_down],
-                                 state: [:link_down],
-                                 curr: [:port_10gb_fd, :port_copper])
+      Pio::OpenFlow10::PhyPort16.new(port_no: 1,
+                                     hardware_address: '11:22:33:44:55:66',
+                                     name: 'port123',
+                                     config: [:port_down],
+                                     state: [:link_down],
+                                     curr: [:port_10gb_fd, :port_copper])
     end
 
     Then { phy_port.port_no == 1 }