+++ spec/pio/open_flow10/set_ip_destination_address_spec.rb
@@ -11,12 +11,12 @@ describe Pio::SetIpDestinationAddress do
         Then { set_ip_destination_addr.ip_address == '1.2.3.4' }
       end
 
-      describe '#type' do
-        Then { set_ip_destination_addr.type == 7 }
+      describe '#action_type' do
+        Then { set_ip_destination_addr.action_type == 7 }
       end
 
-      describe '#message_length' do
-        Then { set_ip_destination_addr.message_length == 8 }
+      describe '#length' do
+        Then { set_ip_destination_addr.length == 8 }
       end
 
       describe '#to_binary' do