+++ spec/pio/open_flow10/set_ip_tos_spec.rb
@@ -9,12 +9,12 @@ describe Pio::SetIpTos do
         Then { set_ip_tos.type_of_service == 32 }
       end
 
-      describe '#type' do
-        Then { set_ip_tos.type == 8 }
+      describe '#action_type' do
+        Then { set_ip_tos.action_type == 8 }
       end
 
-      describe '#message_length' do
-        Then { set_ip_tos.message_length == 8 }
+      describe '#length' do
+        Then { set_ip_tos.length == 8 }
       end
 
       describe '#to_binary' do