+++ spec/pio/open_flow10/set_transport_source_port_spec.rb
@@ -3,22 +3,22 @@ require 'pio/open_flow10/set_transport_port'
 describe Pio::SetTransportSourcePort do
   describe '.new' do
     When(:set_transport_source_port) do
-      Pio::SetTransportSourcePort.new port_number
+      Pio::SetTransportSourcePort.new port
     end
 
     context 'with 100' do
-      When(:port_number) { 100 }
+      When(:port) { 100 }
 
-      describe '#port_number' do
-        Then { set_transport_source_port.port_number == 100 }
+      describe '#port' do
+        Then { set_transport_source_port.port == 100 }
       end
 
-      describe '#type' do
-        Then { set_transport_source_port.type == 9 }
+      describe '#action_type' do
+        Then { set_transport_source_port.action_type == 9 }
       end
 
-      describe '#message_length' do
-        Then { set_transport_source_port.message_length == 8 }
+      describe '#length' do
+        Then { set_transport_source_port.length == 8 }
       end
 
       describe '#to_binary' do
@@ -27,17 +27,17 @@ describe Pio::SetTransportSourcePort do
     end
 
     context 'with -1' do
-      When(:port_number) { -1 }
+      When(:port) { -1 }
       Then { set_transport_source_port == Failure(ArgumentError) }
     end
 
     context 'with 2**16' do
-      When(:port_number) { 2**16 }
+      When(:port) { 2**16 }
       Then { set_transport_source_port == Failure(ArgumentError) }
     end
 
     context 'with :INVALID_PORT_NUMBER' do
-      When(:port_number) { :INVALID_PORT_NUMBER }
+      When(:port) { :INVALID_PORT_NUMBER }
       Then { set_transport_source_port == Failure(TypeError) }
     end
   end